requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-model-utils==4.3.1
djangorestframework==3.14.0
executing==2.0.1
factory-boy==3.3.0
Faker==22.0.0
filelock==3.13.1
flake8==6.1.0
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
iniconfig==2.0.0
ipython==8.20.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pgvector==0.2.4
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.2
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
